7_Auto_Categorizing_News_Artcile/Deployment File/requirements.txt
streamlit==1.11.0
scikit-learn
tensorflow-cpu
pandas==1.3.4
numpy